% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_params.R
\name{beta_params}
\alias{beta_params}
\title{beta_params}
\usage{
beta_params(mean, sd)
}
\arguments{
\item{mean}{Mean of the Beta distribution.}

\item{sd}{Standard deviation of the Beta Distribution.}
}
\value{
The two parameters of the Beta distribution determined by specified mean and sd.
}
\description{
Given mean and sd, this function finds the value of the parameters for a Beta distribution.
}
