% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Z_mc.R
\name{compute_Z_mc}
\alias{compute_Z_mc}
\title{compute_Z_mc
Using the mean and the covariance matrix of a multivariate normal, this function calculates the Monte Carlo estimates simplex probability.}
\usage{
compute_Z_mc(mn_vec, cov_mat, n_mc = 4000, jitter = 1e-10)
}
\arguments{
\item{mn_vec}{The mean vector of the multivariate normal.}

\item{cov_mat}{The covariance matrix of the multivariate normal.}

\item{n_mc}{The number of Monte Carlo samples used to find the estimate. Default is 4000.}

\item{jitter}{A small non-negative value added to the diagonal of the multivariate normal for numerical stabilization. Default is \code{1e-10}.}
}
\value{
A list with two components:
\describe{
  \item{Z}{The Monte Carlo estimate of the simplex probability.}
  \item{se}{The Monte Carlo standard error of the estimate.}
}
}
\description{
compute_Z_mc
Using the mean and the covariance matrix of a multivariate normal, this function calculates the Monte Carlo estimates simplex probability.
}
