% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMGP_data_sim.R
\name{LMGP_data_sim}
\alias{LMGP_data_sim}
\title{LMGP_data_sim}
\usage{
LMGP_data_sim(N, beta, gamma, rho, Tmax, dt = 0.1, iter = 1)
}
\arguments{
\item{N}{Integer. Intial number of susceptible.}

\item{beta}{Numeric. Infection rate parameter.}

\item{gamma}{Numeric. Recovery rate parameter.}

\item{rho}{Numeric. Limiting proportion of initially infected individuals relative to the initially susceptible population. Must satisfy 0 < rho < 1.}

\item{Tmax}{Numeric. Final observation time of epidemic}

\item{dt}{Numeric. time increment. Default is 0.1.}

\item{iter}{Number of dataset wish to simulate.}
}
\description{
Generates a data set of exact infection and recovery times according to the LMGP approximation, where the survival function of time to infection follows a Gaussian process.
}
