% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_data_sim.R
\name{noisy_data_sim}
\alias{noisy_data_sim}
\title{noisy_data_sim}
\usage{
noisy_data_sim(N, beta, gamma, rho, p, Tmax, nu = NULL)
}
\arguments{
\item{N}{Integer. Intial number of susceptible.}

\item{beta}{Numeric. Infection rate parameter.}

\item{gamma}{Numeric. Recovery rate parameter.}

\item{rho}{Numeric. Limiting proportion of initially infected individuals relative to the initially susceptible population. Must satisfy 0 < rho < 1.}

\item{p}{Numeric. The reporting probability parameter. Must satisfy 0 < p < 1.}

\item{Tmax}{Numeric. Final observation time of epidemic.}

\item{nu}{Default is NULL, which simulate usual SIR CTMC dynamics. If specified some numerical value, then, it simulates data from frailty model. It is the parameter denoting the standard deviation of the frailty variable which is assumed to follow Gamma distribution.}
}
\value{
A data set with three columns - days, exact infection counts and under-reported counts with reporting probability \eqn{p}. The true trajectory is simulated via Sellke construction.
}
\description{
Generates a data set with under-reported infection counts simulated via Sellke construction, under-reported by probability \eqn{p}.
}
\seealso{
\code{\link[DSA.CountData]{sellke}} for reference of \eqn{\nu}.
}
