% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sellke_counts.R
\name{sellke_counts}
\alias{sellke_counts}
\title{sellke_counts}
\usage{
sellke_counts(N, beta, gamma, rho, Tmax, nu = NULL)
}
\arguments{
\item{N}{Integer. Intial number of susceptible.}

\item{beta}{Numeric. Infection rate parameter.}

\item{gamma}{Numeric. Recovery rate parameter.}

\item{rho}{Numeric. Limiting proportion of initially infected individuals relative to the initially susceptible population. Must satisfy 0 < rho < 1.}

\item{Tmax}{Numeric. Final observation time of epidemic}

\item{nu}{Default is NULL, which simulate usual SIR CTMC dynamics. If specified some numerical value, then, it simulates data from frailty model. It is the parameter denoting the standard deviation of the frailty variable which is assumed to follow Gamma distribution.}
}
\value{
a data set with two columns - days of infection and the infection counts.
}
\description{
Generates a data set with exact infection counts simulated via Sellke construction.
}
\seealso{
\code{\link[DSA.CountData]{sellke}} for reference of \eqn{\nu}.
}
