% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stan_adj.R
\name{summary_stan_adj}
\alias{summary_stan_adj}
\title{summary_stan_adj
Summarize weighted posterior draws and calculates the posterior coverage and the ESS/s for a single parameter.}
\usage{
summary_stan_adj(
  draws_list,
  wts,
  ess_per_sec_par,
  index,
  true_value,
  not_cap = FALSE
)
}
\arguments{
\item{draws_list}{List of \code{stanfit} objects.}

\item{wts}{List of normalized weights.}

\item{ess_per_sec_par}{Numeric vector containing effective sample size per
 second values for the simulation replicates for a the parameter of interest.
@param index It identifies the parameter we are interested to summarize:  \code{1} = \eqn{\beta}, \code{2} = \eqn{\gamma},
 \code{3} = \eqn{\rho}, and \code{4} = \eqn{R_0}.
 @param true_value The true value of the parameter.}

\item{not_cap}{Logical. If \code{TRUE}, it returns the pairs of 95% CI, which do not contain the true parameter. Otherwise, it returns mean of the means, standard deviations and the proportion of 95% CIs captures the true parameter.}
}
\value{

}
\description{
summary_stan_adj
Summarize weighted posterior draws and calculates the posterior coverage and the ESS/s for a single parameter.
}
\seealso{
\code{\link{summary_stan_adj_list}}
}
