% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stan_adj_list.R
\name{summary_stan_list_adj}
\alias{summary_stan_list_adj}
\title{summary_stan_list_adj}
\usage{
summary_stan_list_adj(result_list, true_value, prob = FALSE, notcap = FALSE)
}
\arguments{
\item{result_list}{List of \code{stanfit} objects that we want to summarize.}

\item{prob}{Logical. If TRUE, it will consider the probability adjusted model and give estimates of \eqn{p}. Default is False.}

\item{true_values}{Vector of true values of the parameters in the order \eqn{\beta}{beta}, \eqn{\gamma}{gamma}, \eqn{\rho}{rho} respectively.}

\item{not_cap}{Logical. If TRUE, it returns the pairs of 95% CI for all four parameters, which do not contain the true parameter. Otherwise, it returns mean of the means, standard deviations and the proportion of 95% CIs captures the true parameter for all four parameters.}
}
\value{

}
\description{
It runs summary_stan_adj functions for all parameters - \eqn{\beta}{beta}, \eqn{\gamma}{gamma}, \eqn{\rho}{rho} (for probability model also \eqn{p}) and returns in the form of a matrix. It can also return the CIs that do not contain the true parameters if notcap = TRUE.
}
\seealso{
\code{\link{summary_stan_adj}}
}
