% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnorm_calc.R
\name{wnorm_calc}
\alias{wnorm_calc}
\title{wnorm_calc
Taking the estimates of estimates of the simplex probabilities, it calculates the normalized weights which are inversely proportional to the probabilities.}
\usage{
wnorm_calc(Z_hat, M)
}
\arguments{
\item{Z_hat}{The vector of Monte-carlo estimates of the simplex probability.}

\item{M}{Positive integer specifying the number of weights to return.}
}
\value{
A list with two components:
\describe{
  \item{ok}{Logical vector indicating which entries of \code{Z_hat} are
  non-missing and positive.}
  \item{w_norm}{Numeric vector of normalized weights. Invalid entries are
  assigned weight zero.}
}
}
\description{
wnorm_calc
Taking the estimates of estimates of the simplex probabilities, it calculates the normalized weights which are inversely proportional to the probabilities.
}
