% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_sd.R
\name{wtd_sd}
\alias{wtd_sd}
\title{wtd_sd
Calculates the weighted standard deriviation}
\usage{
wtd_sd(x, w)
}
\arguments{
\item{x}{vector of values}

\item{w}{vector of sd}
}
\value{
A numeric scalar giving the weighted standard deviation.
}
\description{
wtd_sd
Calculates the weighted standard deriviation
}
\details{
The weighted mean is calculated as
\deqn{\bar{x}_w = \sum_i w_i x_i,}
and the weighted standard deviation is calculated as
\deqn{
\sqrt{\frac{\sum_i w_i(x_i-\bar{x}_w)^2}
{1-\sum_i w_i^2}}.
}
This formula assumes that the weights are normalized so that
\eqn{\sum_i w_i = 1}.
}
