% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMGP_ppc.R
\name{LMGP_ppc}
\alias{LMGP_ppc}
\title{LMGP_ppc}
\usage{
LMGP_ppc(
  fit_full,
  data_obs,
  N,
  prob = TRUE,
  raw_data = NULL,
  w_norm = NULL,
  quantile_probs = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{fit_full}{The stanfit object from the output of LMGP function.}

\item{data_obs}{The final data set (after pre-processing), which was used to fitting. A two-column matrix with first column days and second column is corresponding counts.}

\item{N}{Initial number of susceptible}

\item{raw_data}{A two-column matrix with first column days and second column is raw counts if it differs from the processed data mentioned in \code{data_obs}.}

\item{w_norm}{Normalized weights. If \code{NULL}, equal weights will be used.}

\item{quantile_probs}{A vector of probabilities, will be used to make credible intervals in posterior predictive checks. The default is \code{0.025, 0.5, 0.975}.}
}
\description{
Posterior predictive check for models fitted using LMGP method.
}
